+++ include/ambition/Ambition_vga.hh
@@ -166,6 +166,10 @@ void drawBuildingRallyPoint(
   Town* _7kaaTown
 );
 
+void drawBuildMarkerGridLines(
+  const UserInterface::Rectangle buildMarker
+);
+
 void drawButtonOverlay(
   UserInterface::Rectangle buttonCoordinates,
   std::string text