+++ src/ambition/Ambition_vga.cc
@@ -648,6 +648,75 @@ void drawBuildingRallyPoint(
   }
 }
 
+void drawBuildMarkerGridLines(
+  const UserInterface::Rectangle marker
+) {
+  constexpr auto PIXELS_PER_TILE = 32;
+
+  const auto buildMarker = marker.intersection(UserInterface::VIEWPORT);
+  const auto perimeterMarker
+    = buildMarker.outer(PIXELS_PER_TILE)
+    .intersection(UserInterface::VIEWPORT);
+
+  vga_back.bar(
+    Ambition::UserInterface::VIEWPORT.start.left,
+    perimeterMarker.start.top,
+    Ambition::UserInterface::VIEWPORT.end.left,
+    perimeterMarker.start.top,
+    V_WHITE
+  );
+  vga_back.bar(
+    Ambition::UserInterface::VIEWPORT.start.left,
+    buildMarker.start.top,
+    Ambition::UserInterface::VIEWPORT.end.left,
+    buildMarker.start.top,
+    V_WHITE
+  );
+  vga_back.bar(
+    Ambition::UserInterface::VIEWPORT.start.left,
+    buildMarker.end.top,
+    Ambition::UserInterface::VIEWPORT.end.left,
+    buildMarker.end.top,
+    V_WHITE
+  );
+  vga_back.bar(
+    Ambition::UserInterface::VIEWPORT.start.left,
+    perimeterMarker.end.top,
+    Ambition::UserInterface::VIEWPORT.end.left,
+    perimeterMarker.end.top,
+    V_WHITE
+  );
+
+  vga_back.bar(
+    perimeterMarker.start.left,
+    Ambition::UserInterface::VIEWPORT.start.top,
+    perimeterMarker.start.left,
+    Ambition::UserInterface::VIEWPORT.end.top,
+    V_WHITE
+  );
+  vga_back.bar(
+    buildMarker.start.left,
+    Ambition::UserInterface::VIEWPORT.start.top,
+    buildMarker.start.left,
+    Ambition::UserInterface::VIEWPORT.end.top,
+    V_WHITE
+  );
+  vga_back.bar(
+    buildMarker.end.left,
+    Ambition::UserInterface::VIEWPORT.start.top,
+    buildMarker.end.left,
+    Ambition::UserInterface::VIEWPORT.end.top,
+    V_WHITE
+  );
+  vga_back.bar(
+    perimeterMarker.end.left,
+    Ambition::UserInterface::VIEWPORT.start.top,
+    perimeterMarker.end.left,
+    Ambition::UserInterface::VIEWPORT.end.top,
+    V_WHITE
+  );
+}
+
 void drawButtonOverlay(
   UserInterface::Rectangle buttonCoordinates,
   std::string text