+++ CHANGELOG.md
@@ -12,6 +12,9 @@ speaking, considered to be the player's experience.
 
 ### Added
 
+- When selecting where to place a building, show grid rules of the building, and
+  grid rules one tile out, for a walkable perimeter, to allow for easier
+  placement and planning of buildings.
 - Show the per-unit and total costs, along with the percentages of the fixed
   expense, for units on the Military Report.
 - Show a total for the unit costs, along with the percentage of fixed expense,